Fix IOBEx at infinite inventory usage, close #172

+++ tileentity/GSGraveStoneItems.java
@@ -91,12 +91,12 @@ public class GSGraveStoneItems {
         if (items != null) {
             switch (GraveStoneConfig.graveItemsCount) {
                 case 0:
-                    for (byte i = 0; i < items.size(); i++) {
+                    for (int i = 0; i < items.size(); i++) {
                         dropItem(items.get(i), tileEntity.getWorldObj(), tileEntity.xCoord, tileEntity.yCoord, tileEntity.zCoord);
                     }
                     break;
                 case 40:
-                    for (byte i = 0; i < items.size(); i++) {
+                    for (int i = 0; i < items.size(); i++) {
                         addInventoryContent(items.get(i));
                     }
                     break;
@@ -104,7 +104,7 @@ public class GSGraveStoneItems {
                     int savedItems = GraveStoneConfig.graveItemsCount;
                     Collections.shuffle(Arrays.asList(items.size()), new Random());
 
-                    for (byte i = 0; i < items.size(); i++) {
+                    for (int i = 0; i < items.size(); i++) {
                         if (items.get(i) != null && savedItems > 0) {
                             addInventoryContent(items.get(i));
                             savedItems--;
@@ -119,7 +119,7 @@ public class GSGraveStoneItems {
 
     public void setAdditionalItems(ItemStack[] items) {
         if (items != null) {
-            for (short i = 0; i < items.length; i++) {
+            for (int i = 0; i < items.length; i++) {
                 addInventoryContent(items[i]);
             }
         }
